requirements.txt
kafka-python
langchain-openai
langchain-google-genai
neo4j
python-dotenv
numpy
scikit-learn
sentence-transformers